added creation of scared folder /data
This is needed since each SDcard going into the Tippy Tap has to be
validated as working properly. Since we need to use the init script
anyway we will use the init script copying as a test of the SDcard
destined to each module. that way it makes more sense to create the
data folder when init.sh is used